rename components to parts
