+++ packages/testing/playwright/pages/parts/LogsPanel.ts
